Let Vim decide the keys of the s:prev dictionary.
In order to refrain from doing unnecessary work within the main
function, the s:prev dictionary was populated upon initialization,
mapping the keys to punish against to the value of reltime(). Alas,
special keys such as `<Up>` and `<BS>` undergo an unfortunate
transformation when passed as arguments to functions. This may
well be a bug in Vim, but it's worked around easily enough by
starting with an empty dictionary and letting Vim decide the keys.

This should fix #3.